content/html/document/test/test_bug359657.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=359657
-->
<head>
  <title>Test for Bug 359657</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=359657">Mozilla Bug 359657</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

/** Test for Bug 359657 **/
function runTest() {
  var span = document.createElement("span");
  $("test").insertBefore(span, $("test").firstChild);
  ok(true, "Reachability", "We should get here without crashing");
  is($("test").firstChild, span, "First child is correct");
  SimpleTest.finish();
}
</script>
<div>
  <iframe src=""></iframe>
  <!-- Important: This test needs to run async at this point.  The actual test
       is not crashing while running this test! -->
  <script type="text/javascript" src="data:text/javascript,runTest()"></script>
</div>
</pre>
</body>
</html>
